[REFACTOR] 클래스명 변경, 불필요한 코드 삭제
Squashed commit of the following:

commit 40addc6ece9b404fe114d00d169b66f0fb60075a
Author: conatuseus <conatuseus@gmail.com>
Date:   Mon May 27 13:57:02 2019 +0900

    ICar -> Car,  Car -> CarImpl 로 클래스명 변경
