+++ src/main/java/coordinatecalculator/model/Point.java
@@ -1,7 +1,6 @@
 package coordinatecalculator.model;
 
 
-import java.util.Objects;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
