--- src/main/java/rentcompany/ICar.java
@@ -1,10 +1,0 @@
-package rentcompany;
-
-public interface ICar {
-
-    double getDistancePerLiter();
-
-    double getTripDistance();
-
-    String getName();
-}
